feat:update and remove useless plugins
