+++ .luarc.json
@@ -1,5 +1,6 @@
 {
     "diagnostics.globals": [
-        "Path"
+        "Path",
+        "vim"
     ]
 }
