+++ init.lua
@@ -18,3 +18,5 @@ require("lazy").setup("plugins")
 
 -- require('mods')
 vim.g.NERDTreeWinSize = 42
+
+require("lsp")
